// src/types.ts
type group = "general" | "punctuation" | "arithmetic";
type entitySymbol =
  | "!"
  | "?"
  | "."
  | "@"
  | "#"
  | "$"
  | "%"
  | "^"
  | "&"
  | "("
  | ")"
  | "["
  | "]"
  | "{"
  | "}"
  | "-"
  | "+"
  | "*"
  | "/"
  | "\\"
  | "="
  | "|"
  | "\t"
  | "\b"
  | "\\s"
  | "\n"
  | "`"
  | "~"
  | "_"
  | ","
  | ":"
  | ";"
  | "'"
  | '"'
  | "<"
  | ">";
type name =
  | "exclamationMark"
  | "questionMark"
  | "period"
  | "atSign"
  | "hashTag"
  | "dollarSign"
  | "percent"
  | "caret"
  | "ampersand"
  | "openingParenthesis"
  | "closingParenthesis"
  | "openingBracket"
  | "closingBracket"
  | "openingBrace"
  | "closingBrace"
  | "sum"
  | "plusSign"
  | "difference"
  | "subtract"
  | "product"
  | "multiply"
  | "asterisk"
  | "division"
  | "quotient"
  | "modulus"
  | "percent"
  | "assignment"
  | "verticalBar"
  | "tab"
  | "backSpace"
  | "space"
  | "backtick"
  | "tilde"
  | "underscore"
  | "colon"
  | "semiColon"
  | "singleQuote"
  | "doubleQuote"
  | "apostrophe"
  | "comma"
  | "lessThan"
  | "greaterThan"
  | "newLine"
  | "backSlash"
  | "forwardSlash";

export type EntityGroup = Record<name, HTMLEntity>;
export type HTMLEntity = {
  name: string;
  entityName: string | null;
  symbol: entitySymbol;
  code: string;
  group: group;
};